link_analysis/requirements.txt
joblib
pandas
scikit-learn
pydantic
fastapi
uvicorn
requests
python-dotenv
tldextract
python-whois
